refactor(android): delete LobbyUiState
