--- apps/android/app/src/main/java/at/aau/serg/android/ui/lobby/LobbiesUiState.kt
@@ -1,9 +1,0 @@
-package at.aau.serg.android.ui.lobby
-
-import shared.models.lobby.response.LobbyListItemResponse
-
-sealed class LobbiesUiState {
-    object Loading : LobbiesUiState()
-    data class Success(val lobbies: List<LobbyListItemResponse>) : LobbiesUiState()
-    data class Error(val message: String) : LobbiesUiState()
-}
