--- apps/android/app/src/main/java/at/aau/serg/android/ui/lobby/LobbyUiStateLoading.kt
@@ -1,11 +1,0 @@
-package at.aau.serg.android.ui.lobby
-
-import shared.models.lobby.domain.Lobby
-
-
-sealed class LobbyUiStateLoading {
-    object Loading : LobbyUiStateLoading()
-    data class Success(val lobby: Lobby) : LobbyUiStateLoading()
-    data class Error(val message: String) : LobbyUiStateLoading()
-}
-
